+++ src/library/DIPS.Mobile.UI/Components/Pickers/DatePicker/Inline/iOS/InlineDateAndTimePicker.cs
@@ -1,0 +1,3 @@
+namespace DIPS.Mobile.UI.Components.Pickers.DatePicker.Inline.iOS;
+
+public class InlineDateAndTimePicker : DateAndTimePicker.DateAndTimePicker;