+++ src/library/DIPS.Mobile.UI/Components/Pickers/DatePickerShared/Android/BaseDatePickerHandler.cs
@@ -4,6 +4,7 @@ using DIPS.Mobile.UI.Components.Pickers.DatePicker.Service;
 using DIPS.Mobile.UI.Components.Pickers.TimePicker.Android;
 using Microsoft.Maui.Handlers;
 using Microsoft.Maui.Platform;
+using TimePickerService = DIPS.Mobile.UI.Components.Pickers.TimePicker.TimePickerService;
 
 namespace DIPS.Mobile.UI.Components.Pickers.DatePickerShared.Android;
 
@@ -33,7 +34,7 @@ public class BaseDatePickerHandler : ViewHandler<IDatePicker, Google.Android.Mat
                 DatePickerService.Open(datePicker);
                 break;
             case TimePicker.TimePicker timePicker:
-                TimePickerService.OpenTimePicker(timePicker);
+                TimePickerService.Open(timePicker);
                 break;
         }
     }