+++ src/library/DIPS.Mobile.UI/Components/Pickers/NullableDateAndTimePicker/NullableDateAndTimePicker.cs
@@ -1,56 +1,73 @@
+using DIPS.Mobile.UI.Components.Pickers.DateAndTimePicker;
 using DIPS.Mobile.UI.Components.Pickers.NullableDatePickerShared;
+using DIPS.Mobile.UI.Resources.LocalizedStrings.LocalizedStrings;
+using DIPS.Mobile.UI.Resources.Styles;
+using DIPS.Mobile.UI.Resources.Styles.Chip;
 
 namespace DIPS.Mobile.UI.Components.Pickers.NullableDateAndTimePicker;
 
-public partial class NullableDateAndTimePicker : BaseNullableDatePicker
+public partial class NullableDateAndTimePicker : DateAndTimePicker.DateAndTimePicker, INullableDatePicker
 {
-    private DateAndTimePicker.DateAndTimePicker? m_dateAndTimePicker;
-   
-    protected override View CreateDateOrTimePicker()
+    public NullableDateAndTimePicker()
     {
-        m_dateAndTimePicker = new DateAndTimePicker.DateAndTimePicker
-        {
-            SelectedDateTimeCommand = new Command(OnInternalDateChanged),
-            SelectedDateTime = SelectedDateTime ?? DateTime.Now
-        };
-
-        m_dateAndTimePicker.SetBinding(DatePicker.DatePicker.MinimumDateProperty, new Binding(nameof(MinimumDate), BindingMode.OneWay, source: this));
-        m_dateAndTimePicker.SetBinding(DatePicker.DatePicker.MaximumDateProperty, new Binding(nameof(MaximumDate), BindingMode.OneWay, source: this));
-        m_dateAndTimePicker.SetBinding(DatePicker.DatePicker.IgnoreLocalTimeProperty, new Binding(nameof(IgnoreLocalTime), BindingMode.OneWay, source: this));
-        return m_dateAndTimePicker;
-    }
-
-    protected override bool IsDateOrTimeNull()
-    {
-        return SelectedDateTime is null;
+        DateChip.CloseCommand = new Command(OnCloseTapped);
+        TimeChip.CloseCommand = new Command(OnCloseTapped);
     }
     
-    private void OnInternalDateChanged()
+    protected override void OnHandlerChanged()
     {
-        SelectedDateTime = m_dateAndTimePicker?.SelectedDateTime;
-    }
-    
-    protected override void OnSwitchToggled(object? sender, ToggledEventArgs e)
-    {
-        base.OnSwitchToggled(sender, e);
+        base.OnHandlerChanged();
         
-        if (e.Value)
-        {
-            OnInternalDateChanged();
-        }
-        else
-        {
-            SelectedDateTime = null;
-        }
+        OnSelectedDateTimeChanged();
+    }
+
+    public override DateTimeKind GetDateTimeKind()
+    {
+        return SelectedDateTime?.Kind ?? DateTimeKind.Unspecified;
     }
 
     private void OnSelectedDateTimeChanged()
     {
-        if (m_dateAndTimePicker is not null)
-            m_dateAndTimePicker.SelectedDateTime = SelectedDateTime ?? m_dateAndTimePicker.SelectedDateTime;
-        else return;
-        
-        DateEnabledSwitch.IsToggled = SelectedDateTime.HasValue;
-        SelectedDateTimeCommand?.Execute(null);
+        if (SelectedDateTime is null)
+        {
+            DateChip.IsCloseable = false;
+            TimeChip.IsCloseable = false;
+            
+            DateChip.Style = Styles.GetChipStyle(ChipStyle.EmptyInput);
+            TimeChip.Style = Styles.GetChipStyle(ChipStyle.EmptyInput);
+
+            DateChip.Title = DUILocalizedStrings.ChooseDate;
+            TimeChip.Title = DUILocalizedStrings.ChooseTime;
+        }
+        else
+        {
+            DateChip.IsCloseable = true;
+            TimeChip.IsCloseable = true;
+            
+            DateChip.Style = Styles.GetChipStyle(ChipStyle.Input);
+            TimeChip.Style = Styles.GetChipStyle(ChipStyle.Input);
+            
+            OnSelectedDateTimeChanged(SelectedDateTime.Value);
+        }
     }
+
+    public override void SetSelectedDateTime(DateTime? selectedDate)
+    {
+        base.SetSelectedDateTime(selectedDate);
+
+        SelectedDateTime = selectedDate is null ? null : ValidateDateTime(selectedDate.Value);
+    }
+
+    public override DateTime GetDateOnOpen()
+    {
+        return ValidateDateTime(SelectedDateTime ?? DateTime.Now);
+    }
+    
+    private void OnCloseTapped()
+    {
+        SelectedDateTime = null;
+
+        DateAndTimePickerService.Close();
+    }
+
 }