+++ src/library/DIPS.Mobile.UI/Components/Pickers/TimePicker/Android/TimePickerService.cs
@@ -1,11 +1,13 @@
-namespace DIPS.Mobile.UI.Components.Pickers.TimePicker.Android;
+using DIPS.Mobile.UI.Components.Pickers.TimePicker.Android;
 
-public class TimePickerService
+namespace DIPS.Mobile.UI.Components.Pickers.TimePicker;
+
+public partial class TimePickerService
 {
     private static MaterialTimePickerFragment? m_materialDateTimePicker;
     internal const string TimePickerTag = "DUIMaterialTimePicker";
 
-    public static void OpenTimePicker(TimePicker timePicker)
+    public static partial void Open(TimePicker timePicker, View? sourceView = null)
     {
         if (IsOpen())
             Close();
@@ -13,9 +15,9 @@ public class TimePickerService
         m_materialDateTimePicker = new MaterialTimePickerFragment(timePicker);
     }
 
-    internal static bool IsOpen() => m_materialDateTimePicker != null && m_materialDateTimePicker.IsOpen();
+    internal static partial bool IsOpen() => m_materialDateTimePicker != null && m_materialDateTimePicker.IsOpen();
 
-    public static void Close()
+    public static partial void Close()
     {
         m_materialDateTimePicker?.Close();
     }